refactor: move `routes` to its own package
